refactor: alterando style do componente logout - @vanessalmirante @MarquesEmilly2
