+++ src/components/logout/styled.js
