+++ src/pages/home/index.js
@@ -3,15 +3,15 @@ import { faChevronCircleRight } from '@fortawesome/free-solid-svg-icons'
 import { HomePageContainer, HomeContainer } from './styled'
 import { Hyperlink } from '../../components/hyperlink'
 import { isAdmin } from '../../utils/isAdmin'
-import { Logout } from '../../components/layout'
+import { Logout } from '../../components/logout'
 
 export const HomePage = () => {
   const { t } = useTranslation()
   return (
     <HomePageContainer>
       <div className='header'>
-        <h1>{t('home.title')}</h1>
         <Logout />
+        <h1>{t('home.title')}</h1>
         <HomeContainer>
           <Hyperlink link="/hiring-process" label={t('hiringProcess.title')} icon={faChevronCircleRight} />
           {isAdmin() &&
